View the summary of which files were affected... Big change:    In the code the players name was changed from beaver to player. Likewise isBeaver was changed to isPlayer.
An attempt was made to make the character generation separate from the level. So that generating the character could be done as player = beaverChar(), after calling require("Scripts.characters"). This was done to facilitate multiple characters. The attempt appears successful so far except that it appears the physics engine is blatantly ignoring the angularDamping, which the code clearly demonstrates has been set correctly. This is currently a new bug, but I consider it small for the trade up of being able to generate the characters in a much cleaner way.

level2/level2b still needs to be updated, including their build files which contain unneccessary character generation code in the asset section as well as below where the sprite is defined.