Chris: Credits Scene completly redone.
The code is actually quite sloppy, but it's very easy to use and the result is quite decent looking I think.

The timing for changing the scene back to the main menu isn't automatic yet, but it can be made that way. As a result, if you end the scene too soon not all of the display objects get removed from the scene and are stuck in the display permanently, apparently.

each creditsObject has 1 image (our face) and it needs to be 200x200 pixels