I modified the code so that the program would run and compile with the new version of Corona.
This required explicitly declaring some functions as global functions, where as previously that was implicit.

I also downloaded the sprite.lua library which we used in the previous version of Corona, but is now currently depreciated.

I also had to enable a new compatibility library that tells Corona to support the old scene Director, which is now currently depreciated.

This code has been compiled and tested on android. It appears to work as well as it did before the new version of Corona was released.